memory_agent/requirement.txt
aiofiles==24.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.10.11
aiolimiter==1.2.1
aiomysql==0.2.0
aiosignal==1.3.2
alembic==1.13.2
anaconda-anon-usage @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_60q98_n2ty/croot/anaconda-anon-usage_1732732446690/work
annotated-types==0.7.0
anyio==4.5.0
archspec @ file:///croot/archspec_1709217642129/work
attrs==25.1.0
backoff==2.2.1
bcrypt==4.2.0
bidict==0.23.1
boltons @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_0a8f8rf2dq/croot/boltons_1737061714423/work
boto3==1.40.69
botocore==1.40.69
Brotli @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_f7i0oxypt6/croot/brotli-split_1736182464088/work
certifi @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_53iribkvo4/croot/certifi_1738623743875/work/certifi
cffi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_e4xd9yd9i2/croot/cffi_1736182819442/work
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
conda @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_f7tjb4b3kt/croot/conda_1743715301936/work/conda-src
conda-anaconda-telemetry @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_70bir5mxml/croot/conda-anaconda-telemetry_1736524610507/work
conda-anaconda-tos @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_a5mn_e8dc_/croot/conda-anaconda-tos_1739299138582/work
conda-content-trust @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_9fusbfzixa/croot/conda-content-trust_1714483157715/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1737733694612/work/src
conda-package-handling @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_82jv6q9ech/croot/conda-package-handling_1731369032774/work
conda_package_streaming @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_2el5nlypud/croot/conda-package-streaming_1731366187356/work
croniter==6.0.0
cryptography==43.0.1
dataclasses-json==0.6.7
distro @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_ddkyz0575y/croot/distro_1714488254309/work
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
filelock==3.17.0
frozendict @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_8b0cest_id/croot/frozendict_1713194839836/work
frozenlist==1.5.0
fsspec==2024.12.0
grandalf==0.8
greenlet==3.1.0
grpcio==1.76.0
h11==0.14.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.4.2
jose==1.0.0
json_repair==0.44.1
jsonpatch @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_3ajyoz8zoj/croot/jsonpatch_1714483362270/work
jsonpointer==3.0.0
langchain==1.1.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.2
langchain-neo4j==0.6.0
langchain-openai==0.3.8
langchain-pinecone==0.2.3
langchain-tests==0.3.14
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==2.1.0
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.14
langsmith==0.3.45
libmambapy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_8fxbc14zt9/croot/mamba-split_1734469521691/work/libmambapy
llvmlite==0.44.0
loguru==0.7.2
lxml==5.3.1
Mako==1.3.5
markdown-it-py @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/markdown-it-py_1699237863445/work
MarkupSafe==2.1.5
marshmallow==3.26.1
mdurl @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/mdurl_1699237660008/work
mem0ai==1.0.1
menuinst @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_66is8p984e/croot/menuinst_1738943433761/work
more-itertools==10.6.0
motor==3.5.3
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
neo4j==5.28.2
neo4j-graphrag==1.11.0
networkx==3.4.2
numba==0.61.0
numpy==2.4.0
openai==2.14.0
openai-whisper==20240930
orjson==3.10.15
ormsgpack==1.10.0
packaging @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_a6_qk3qyg7/croot/packaging_1734472142254/work
pandas==2.2.3
pillow==11.1.0
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
platformdirs @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/platformdirs_1701803010714/work
pluggy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_bei6q1cvvt/croot/pluggy_1733169621717/work
portalocker==3.2.0
posthog==7.4.0
propcache==0.3.0
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pycosat @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_61k1o_oje7/croot/pycosat_1736868429741/work
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.10.5
pydantic-settings==2.12.0
pydantic_core==2.23.4
pydub==0.25.1
Pygments @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/pygments_1699238063295/work
pymongo==4.8.0
PyMuPDF==1.25.3
PyMySQL==1.1.1
pyparsing==3.2.3
pypdf==6.3.0
PySocks @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/pysocks_1699237568675/work
pytesseract==0.3.13
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.11.2
python-jose==3.3.0
python-multipart==0.0.9
python-socketio==5.9.0
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.16.2
rank-bm25==0.2.2
redis==7.1.0
regex==2024.11.6
requests==2.32.5
requests-toolbelt==1.0.0
rich @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_034myiu7pr/croot/rich_1732638981241/work
rich-toolkit==0.14.8
rignore==0.6.2
rq==2.6.0
rsa==4.9
ruamel.yaml @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_35yvtl3p84/croot/ruamel.yaml_1727980165481/work
ruamel.yaml.clib @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_70l_orv46q/croot/ruamel.yaml.clib_1727769819918/work
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.15.2
sentry-sdk==2.11.0
setuptools==75.8.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
sympy==1.13.1
syrupy==4.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.22.1
torch==2.6.0
torchaudio==2.6.0
tqdm @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_5642tzrprr/croot/tqdm_1738943467095/work
transformers==4.57.1
truststore @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_a1qa3ogm4o/croot/truststore_1736550138575/work
typer==0.16.0
types-PyYAML==6.0.12.20250915
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_0b3jpv_f79/croot/typing_extensions_1734714864260/work
tzdata==2025.1
ujson==5.10.0
urllib3 @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_8dwi5l2dj0/croot/urllib3_1737133640453/work
uuid_utils==0.12.0
uvicorn==0.22.0
uvloop==0.21.0
watchfiles==1.1.0
watchtower==3.1.0
websockets==15.0.1
wheel==0.45.1
win32-setctime==1.1.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
zstandard @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_a5_i6g4o6n/croot/zstandard_1731356352787/work
